pyproject.toml
[tool.poetry]
name = "fast-text-test"
version = "0.1.0"
description = ""
authors = ["KiharaTakahiro <takahirokihara123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
mecab-python3 = "^1.0.10"
ipadic = "^1.0.0"
fasttext-wheel = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
